refactor: FairyLevelPassiveSkillCondition

+++ src/main/kotlin/miragefairy2023/modules/passiveskill/passiveSkillConditions.kt
@@ -289,7 +289,7 @@ class StatusEffectPassiveSkillCondition(private val statusEffect: StatusEffect) 
     override fun test(player: PlayerEntity, itemStack: ItemStack) = player.hasStatusEffect(statusEffect)
 }
 
-class FairyLevelPassiveSkillCondition(private val fairyLevel: Int) : PassiveSkillCondition {
+class FairyLevelPassiveSkillCondition(private val fairyLevel: Double) : PassiveSkillCondition {
     override fun getText() = text { "${Symbol.STAR}$fairyLevel${Symbol.UP}"() }
     override fun test(player: PlayerEntity, itemStack: ItemStack) = itemStack.item.castOr<DemonFairyItem> { return false }.fairyLevel >= fairyLevel
 }
